Fixed phaser and Green3 Level, fixed player aiming and attempted mushroom fix, fixed spear interactions with fruits and other objects

+++ Assets/Scripts/PlayerAiming.cs
@@ -8,7 +8,7 @@ public class PlayerAiming : MonoBehaviour
 {
 
     //public variables
-    public int targetDisplacement = 5;
+    public float targetDisplacement = 5;
     public float upDist = 2;
     public float downDist = 1;
     public float heightBuffer = 0.5f;
